Move platform specifics out of Solaredge const
